refactor: 편집 상태에 따른 컴포넌트 분리 및 리팩토링
